feat: 공휴일 스케줄 검증 로직 추가

+++ src/main/java/com/groupMeeting/holiday/schedule/HolidayScheduler.java
@@ -48,6 +48,10 @@ public class HolidayScheduler {
                         .toInstant()
         );
 
+        if (holiday == null || holiday.events().isEmpty()) {
+            return;
+        }
+
         String date = today.toString();
         holidayRepository.deleteAll(holidayRepository.getHolidayByYearAndMonth(date.substring(0, 4), date.substring(5, 7)));
